fix(onboard): extract host proxy env helper
